Fix a bug when positioning within big compact files
git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/goby@25549 60cd2ccc-2b05-0410-9566-da273a940bff

+++ src/edu/cornell/med/icb/goby/reads/FastBufferedMessageChunksReader.java
@@ -79,7 +79,7 @@ public class FastBufferedMessageChunksReader extends MessageChunksReader {
 
         int b;
         int contiguousZeroBytes = 0;
-        int skipped = 0;
+        long skipped = 0;
         long position = 0;
 
         // search though the input stream until a delimiter chunk or end of stream is reached
